backend/requirements.txt
# Django Core
Django==4.2.25
gunicorn==22.0.*
django-ninja==1.2.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.11
python-decouple==3.8
django-cors-headers==4.9.0
Pillow==11.3.0
django-extensions==4.1
django-filter==24.3
djangorestframework-simplejwt==5.3.0
django-allauth==65.11.2
django-redis==6.0.0

# Storage & External Services
boto3==1.35.*
requests==2.32.*
python-dotenv==1.0.*
openai>=1.0.0
pgvector>=0.2.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0

# Analytics
user-agents>=2.2.0

# Development tools (optional)
django-debug-toolbar==4.4.6